Tambah Policy

+++ app/Http/Controllers/BukuController.php
@@ -8,11 +8,17 @@ use App\Models\Author;
 use App\Models\Penerbit;
 use App\Models\Tahun;
 use App\Models\Genre;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\DB;
+
 class BukuController extends Controller
 {
+    use AuthorizesRequests;
     public function index(Request $request)
     {
-        $query=Buku::with(['authors','penerbit','tahun','genre'])->where('user_id',auth()->id());
+        $this->authorize('viewAny',Buku::class);
+        $query=Buku::with(['authors','penerbit','tahun','genre'])->where('user_id',Auth::id());
         if($request->has('search') && $request->search !=''){
             $query->where('judul','like','%'.$request->search.'%');
         }
@@ -21,15 +27,17 @@ class BukuController extends Controller
     }
     public function create()
     {
-        $authors=Author::where('user_id',auth()->id())->get();
-        $penerbits=Penerbit::where('user_id',auth()->id())->get();
-        $tahuns=Tahun::where('user_id',auth()->id())->get();
-        $gen=Genre::where('user_id',auth()->id())->get();
+        $this->authorize('create',Buku::class);
+        $authors=Author::where('user_id',Auth::id())->get();
+        $penerbits=Penerbit::where('user_id',Auth::id())->get();
+        $tahuns=Tahun::where('user_id',Auth::id())->get();
+        $gen=Genre::where('user_id',Auth::id())->get();
 
         return view('main.create',compact(['authors','penerbits','tahuns','gen']));
     }
     public function store(Request $request)
     {
+        $this->authorize('create',Buku::class);
         $request->validate([
             'judul'=>'required|string',
             'author_id'=>'required|array',
@@ -47,7 +55,7 @@ class BukuController extends Controller
         }
         $bukus=Buku::create([
             'judul'=>$request->judul,
-            'user_id'=>auth()->id(),
+            'user_id'=>Auth::id(),
             'penerbit_id'=>$request->penerbit_id,
             'tahun_id'=>$request->tahun_id,
             'genre_id'=>$request->genre_id,
@@ -59,15 +67,17 @@ class BukuController extends Controller
     public function show(string $id)
     {
         $bukus=Buku::findOrFail($id);
+        $this->authorize('view',$bukus);
         return view('main.show',compact('bukus'));
     }
     public function edit(string $id)
     {
         $bukus=Buku::findOrFail($id);
-        $authors=Author::where('user_id',auth()->id())->get();
-        $penerbits=Penerbit::where('user_id',auth()->id())->get();
-        $tahuns=Tahun::where('user_id',auth()->id())->get();
-        $gen=Genre::where('user_id',auth()->id())->get();
+        $this->authorize('update',$bukus);
+        $authors=Author::where('user_id',Auth::id())->get();
+        $penerbits=Penerbit::where('user_id',Auth::id())->get();
+        $tahuns=Tahun::where('user_id',Auth::id())->get();
+        $gen=Genre::where('user_id',Auth::id())->get();
         return view('main.edit',compact(['bukus','authors','penerbits','tahuns','gen']));
     }
     public function update(Request $request, string $id)
@@ -81,6 +91,7 @@ class BukuController extends Controller
             'genre_id'=>'required|exists:genres,id',
         ]);
         $bukus=Buku::findOrFail($id);
+        $this->authorize('update',$bukus);
         if($request->hasFile('gambar')){
             $ext=$request->file('gambar')->getClientOriginalExtension();
             $namaFile=now()->format('YmdHis').'.'.$ext;
@@ -99,7 +110,18 @@ class BukuController extends Controller
     public function destroy(string $id)
     {
         $bukus=Buku::findOrFail($id);
+        $this->authorize('delete',$bukus);
         $bukus->delete();
         return redirect()->route('daftar.buku')->with('success',$bukus->judul.' Berhasil dihapus');
+    }
+    public function searchVulnerable(Request $request){
+        $search=$request->search;
+        $bukus=DB::select("select*from bukus where judul like '%$search%'");
+        return response()->json($bukus);
+    }
+    public function safeSearch(Request $request){
+        $search=$request->search;
+        $bukus=DB::where('judul','like','%'.$search.'%')->where('user_id',Auth::id())->get();
+        return response()->json($bukus);
     }
 }
